requirements.txt
pytest >= 7.3.0
requests >= 2.28.2
python-dotenv >= 1.0.0